+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fix `setAngle` for different originX/originY (!= 'center')
 - Change default/init noise/brightness value for `fabric.Image.filters.Noise` and `fabric.Image.filters.Brightness` from 100 to 0
 - Add `fabric.Canvas#imageSmoothingEnabled`
+- Add `copy/paste` support for iText (uses clipboardData)
 
 **Version 1.4.0**